+++ test/src/analyzers/lint_analyzer/metrics/metrics_list/technical_debt/technical_debt_metric_test.dart
@@ -22,8 +22,8 @@ Future<void> main() async {
             'ignore-cost': 320,
             'ignore-for-file-cost': 396,
             'as-dynamic-cost': 322,
-            'deprecated-annotations': 37,
-            'file-nullsafety-migration': 41,
+            'deprecated-annotations-cost': 37,
+            'file-nullsafety-migration-cost': 41,
           },
         },
       );