+++ test/src/cli/commands/check_unused_l10n_command_test.dart
@@ -23,6 +23,9 @@ const _usage = 'Check unused localization in *.dart files.\n'
     '    --exclude=<{/**.g.dart,/**.template.dart}>    File paths in Glob syntax to be exclude.\n'
     '                                                  (defaults to "{/**.g.dart,/**.template.dart}")\n'
     '\n'
+    '\n'
+    '    --[no-]fatal-unused                           Treat find unused l10n as fatal.\n'
+    '\n'
     'Run "metrics help" to see global options.';
 
 void main() {